add range to track control
