+++ untitled folder/Atm_wav_1 2.h
@@ -10,6 +10,8 @@ added play function has input filename
 added instances, this version with 2 instances
 looping works
 8/1/17
+rewrite for fewer states
+8/26/17
 */
 
 #include <Automaton.h>
@@ -22,29 +24,31 @@ looping works
 
 class Atm_wav_1 : public Machine {
 	public:
-		enum { WAV_1_ON, WAV_2_ON, WAV_3_ON, WAV_4_ON, WAV_5_ON, WAV_OFF }; //STATES
-		enum { EVT_ENC_DOWN, EVT_ENC_UP, EVT_BTN_1, EVT_PLAY_CHECK, ELSE }; //EVENTS
+		enum {  WAV_PLAY, WAV_OFF, TRACK_UP, TRACK_DOWN, BTN_1, /*PLAY_CHECK*/ }; //STATES
+		enum { EVT_WAV_OFF, EVT_WAV_PLAY, EVT_ENC_DOWN, EVT_ENC_UP, EVT_BTN_1, EVT_PLAY_CHECK, ELSE }; //EVENTS
 		//enum {WAV_START, WAV_STOP}; //ACTIONS
 		
 		Atm_wav_1( void ) : Machine(){};
 		Atm_wav_1& begin( int instance );
 		Atm_wav_1& play( const char *filename ); 
 		Atm_wav_1& stop( void );
+		Atm_wav_1& encoderUp( void );
+		Atm_wav_1& encoderDown( void );
+		Atm_wav_1& btn1( void );
 		Atm_wav_1& playCheck( void );
 		Atm_wav_1& trace( Stream& stream );
-		Atm_wav_1& onPress( Machine& machine, int event = 0 );
-		Atm_wav_1& onPress( atm_cb_push_t callback, int idx = 0 ); 
+		//Atm_wav_1& onPress( Machine& machine, int event = 0 );
+		//Atm_wav_1& onPress( atm_cb_push_t callback, int idx = 0 ); 
 		Atm_wav_1& trigger( int event );
 		int state( void );
 		
 		
 	private:
-     	enum { ENT_WAV_1_ON, ENT_WAV_2_ON, ENT_WAV_3_ON, ENT_WAV_4_ON, ENT_WAV_5_ON, ENT_WAV_OFF, ENT_PLAY_CHECK }; // ACTIONS
-     	enum { ON_PRESS, CONN_MAX }; // CONNECTORS
-     	atm_connector connectors[CONN_MAX];
-		int playerNo;
-		//atm_timer_millis check_timer;
-		//atm_counter counter;
+     	enum { ENT_WAV_PLAY, ENT_WAV_OFF, ENT_TRACK_UP, ENT_TRACK_DOWN, ENT_BTN_1, ENT_PLAY_CHECK }; // ACTIONS
+     	//enum { ON_PRESS, CONN_MAX }; // CONNECTORS
+     	//atm_connector connectors[CONN_MAX];
+		int player_instance;
+		int display_delay;
 
 		
      	int event( int id );
